further specialize transactional queues and hubs
